+++ src/scripts/main.ts
@@ -2,17 +2,18 @@ import {LeaDocumentsOverview} from './components/documents-overview/overview';
 import {LeaDocumentsContainer} from './components/documents-overview/container';
 import {injectOmniplusLogo} from './components/logo-patcher';
 import {injectDocumentsOverviewButtonToLea} from './components/documents-overview/overview-button';
-import {removeAllLineBreaks, removeHeaderImage, removeLeaAnchorHoverCSSRule} from './components/page-patcher';
-import {ForumMessage} from './components/forum-page/forum-message';
+import {
+    autoLogin,
+    removeAllLineBreaks,
+    removeHeaderImage,
+    removeLeaAnchorHoverCSSRule
+} from './components/page-patcher';
 import {ForumSubject} from './components/forum-page/forum-subject';
 
 require('./omniplus.css');
 
 removeHeaderImage();
 injectOmniplusLogo();
-
-// @ts-ignore
-window['ForumMessage'] = ForumMessage;
 
 // If the script is being run on Lea.
 if (window.location.href.includes('ovx.omnivox.ca')) {
@@ -36,5 +37,12 @@ if (window.location.href.includes('ovx.omnivox.ca')) {
             subject.injectToForumSubjectPage();
             subject.render();
         }
+    }
+}
+// If the script is being run on Omnivox/MIO
+else {
+    // If this is the login page.
+    if (window.location.href.includes('Login')) {
+        autoLogin();
     }
 }